#43 problem: 229. Majority Element  and solve previous mistake
