+++ Medium/287. Find Duplicate Number/sol.cpp
@@ -1,28 +1,32 @@
 #include <iostream>
 #include <vector>
-#include <map> 
 
 using namespace std;
 
 class Solution
 {
 public:
-    vector<int> majorityElement(vector<int>& nums) {
-        vector<int> vec;
-        int n= floor(nums.size()/3)+1;
-        map<int,int> mp;
-        map<int,int>::iterator it;    
-        for(int i=0;i<nums.size();i++){
-            it=mp.find(nums[i]);
-            if(it==mp.end())        
-                mp.insert(pair<int,int>(nums[i],1));        
-            else it->second=mp[nums[i]]+1;        
-        } 
-        for(it=mp.begin();it!=mp.end();it++){
-            if(it->second>=n)        
-                vec.push_back(it->first);        
+    int findDuplicate(vector<int> &nums)
+    {
+
+        int slow = nums[0];
+        int fast = nums[0];
+
+        do
+        {
+            slow = nums[slow];
+            fast = nums[nums[fast]];
+        } while (slow != fast);
+
+        fast = nums[0];
+
+        while (slow != fast)
+        {
+            slow = nums[slow];
+            fast = nums[fast];
         }
-        return vec;   
+
+        return slow;
     }
 };
